Move out renderer to separate module
